[Edit] Edited UserInfo.js and ChallengeTable.js to connect with db through axios
[Edit] Edited UserPage.js to have a better modularity of code calling UserInfo
